refactor(registries): move KnownRegistries out of the constants package
`@pnpm/constants` holds constants. `createKnownRegistries` is a factory that
parses URLs, normalizes paths, sorts, and memoizes — none of which the
package name leads a reader to expect. `BUILTIN_NAMED_REGISTRIES` stays,
since that genuinely is one.

`@pnpm/config.pick-registry-for-package` is the natural home: its whole job
is deciding which registry a package comes from, and the known-registry set
is the input to that decision. Scope-based selection (`pickRegistryForPackage`)
and alias-based selection now sit together, which also matches pacquet, where
`KnownRegistries` already lives beside `pick_registry_for_package` and
`pick_registry_for_version` in `named_registry.rs`.

The pinning test moves with the code it covers.
