+++ pnpm11/config/pick-registry-for-package/package.json
@@ -31,6 +31,7 @@
     ".test": "cross-env NODE_OPTIONS=\"$NODE_OPTIONS --experimental-vm-modules --disable-warning=ExperimentalWarning --disable-warning=DEP0169\" jest"
   },
   "dependencies": {
+    "@pnpm/constants": "workspace:*",
     "@pnpm/types": "workspace:*"
   },
   "devDependencies": {
