+++ pnpm11/config/pick-registry-for-package/tsconfig.json
@@ -10,6 +10,9 @@
   ],
   "references": [
     {
+      "path": "../../core/constants"
+    },
+    {
       "path": "../../core/types"
     }
   ]
