+++ pnpm11/deps/compliance/sbom/package.json
@@ -35,6 +35,7 @@
   },
   "dependencies": {
     "@cyclonedx/cyclonedx-library": "catalog:",
+    "@pnpm/config.pick-registry-for-package": "workspace:*",
     "@pnpm/constants": "workspace:*",
     "@pnpm/deps.compliance.license-resolver": "workspace:*",
     "@pnpm/error": "workspace:*",
