+++ pnpm11/installing/deps-resolver/tsconfig.json
@@ -16,6 +16,9 @@
       "path": "../../catalogs/types"
     },
     {
+      "path": "../../config/pick-registry-for-package"
+    },
+    {
       "path": "../../config/version-policy"
     },
     {
