+++ pnpm11/lockfile/utils/tsconfig.json
@@ -10,6 +10,9 @@
   ],
   "references": [
     {
+      "path": "../../config/pick-registry-for-package"
+    },
+    {
       "path": "../../core/constants"
     },
     {
